Properly enable rack_cache to support X-Accel-Redirect
There is an issue where if one only turns on rack-cache and doesn't configure it to use file based caching, then it won't pass a `body` up the middleware stack that responds to `to_path`. If that happens, `rack-sendfile` won't invoke, and all X-Sendfile / X-Accel-Redirect functionality will silently fail. See:

https://github.com/rack/rack/blob/master/lib/rack/sendfile.rb#L112

I'm pretty sure this applies to Rails 3.1 and 3.2 but can't test it easily so I've left that part as-is.